+++ .dockerignore
@@ -1,0 +1,25 @@
+## folders
+.venv/
+.env/
+__pycache__/
+*/__pycache__/
+.pytest_cache/
+*/.pytest_cache/
+.mypy_cache/
+
+## files
+*.pyc
+*.log
+my.db
+*.log.zip
+my_install.sh
+.python-version
+.pre-commit-config.yaml
+
+# dotenv
+.env
+
+# test file
+test.py
+*.session
+*.session-journal
