+++ bot/deps/main.py
@@ -2,7 +2,7 @@ from dependency_injector import containers
 from dependency_injector import providers
 from pyrogram.client import Client as TelegramClient
 
-from src.config import settings
+from bot.config import settings
 
 
 class MainContainer(containers.DeclarativeContainer):
