+++ bot/telegram/commands/kas.py
@@ -6,16 +6,16 @@ from pyrogram.types import InlineKeyboardButton
 from pyrogram.types import InlineKeyboardMarkup
 from pyrogram.types import Message
 
-from src.config import constants
-from src.config import photo_ids
-from src.telegram.base import PrivateCommandEndpoint
-from src.telegram.base.callback_query import CallbackQueryEndpoint
-from src.telegram.templates import text as templates_text
+from bot.config import constants
+from bot.config import photo_ids
+from bot.telegram.base import CallbackQueryEndpoint
+from bot.telegram.base import PrivateCommandEndpoint
+from bot.telegram.templates import text as templates_text
 
 
-class FethiyeEndpoint(PrivateCommandEndpoint):
+class KasEndpoint(PrivateCommandEndpoint):
     commands: List[str] = [
-        constants.SOUTH,
+        constants.KAS,
     ]
 
     async def handle(
@@ -25,9 +25,9 @@ class FethiyeEndpoint(PrivateCommandEndpoint):
     ) -> None:  # noqa: U100
         await client.send_photo(
             chat_id=message.from_user.id,
-            photo=photo_ids.SOUTH,
+            photo=photo_ids.KAS,
         )
-        template = templates_text.SOUTH
+        template = templates_text.KAS
         await client.send_message(
             chat_id=message.from_user.id,
             text=template,
@@ -35,7 +35,7 @@ class FethiyeEndpoint(PrivateCommandEndpoint):
                 [
                     [
                         InlineKeyboardButton(
-                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.SOUTH}"
+                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.KAS}"
                         ),
                         InlineKeyboardButton(text=constants.MAIN_MENU, callback_data=constants.MENU),
                     ]
@@ -44,8 +44,8 @@ class FethiyeEndpoint(PrivateCommandEndpoint):
         )
 
 
-class FethiyeCallbackQueryEndpoint(CallbackQueryEndpoint):
-    callback_query_name: str = constants.SOUTH
+class KasCallbackQueryEndpoint(CallbackQueryEndpoint):
+    callback_query_name: str = constants.KAS
 
     async def handle(
         self,
@@ -53,18 +53,18 @@ class FethiyeCallbackQueryEndpoint(CallbackQueryEndpoint):
         callback_query: CallbackQuery,
     ) -> None:  # noqa: U100
         await client.answer_callback_query(callback_query.id)
+        template = templates_text.KAS
         await client.send_photo(
             chat_id=callback_query.from_user.id,
-            photo=photo_ids.SOUTH,
+            photo=photo_ids.KAS,
         )
-        template = templates_text.SOUTH
         await callback_query.message.reply(
             text=template,
             reply_markup=InlineKeyboardMarkup(
                 [
                     [
                         InlineKeyboardButton(
-                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.SOUTH}"
+                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.KAS}"
                         ),
                         InlineKeyboardButton(text=constants.MAIN_MENU, callback_data=constants.MENU),
                     ]
