+++ bot/telegram/commands/south.py
@@ -6,17 +6,16 @@ from pyrogram.types import InlineKeyboardButton
 from pyrogram.types import InlineKeyboardMarkup
 from pyrogram.types import Message
 
-from src.config import constants
-from src.config import photo_ids
-from src.telegram.base import CallbackQueryEndpoint
-from src.telegram.base import PrivateCommandEndpoint
-from src.telegram.templates import text as templates_text
+from bot.config import constants
+from bot.config import photo_ids
+from bot.telegram.base import PrivateCommandEndpoint
+from bot.telegram.base.callback_query import CallbackQueryEndpoint
+from bot.telegram.templates import text as templates_text
 
 
-class StartEndpoint(PrivateCommandEndpoint):
+class FethiyeEndpoint(PrivateCommandEndpoint):
     commands: List[str] = [
-        constants.START,
-        constants.MENU,
+        constants.SOUTH,
     ]
 
     async def handle(
@@ -26,25 +25,27 @@ class StartEndpoint(PrivateCommandEndpoint):
     ) -> None:  # noqa: U100
         await client.send_photo(
             chat_id=message.from_user.id,
-            photo=photo_ids.START,
+            photo=photo_ids.SOUTH,
         )
-        template = templates_text.MENU
+        template = templates_text.SOUTH
         await client.send_message(
             chat_id=message.from_user.id,
             text=template,
             reply_markup=InlineKeyboardMarkup(
                 [
                     [
-                        InlineKeyboardButton(text=constants.GUIDE_SOUTH, callback_data=constants.SOUTH),
-                        InlineKeyboardButton(text=constants.GUIDE_KAS, callback_data=constants.KAS),
+                        InlineKeyboardButton(
+                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.SOUTH}"
+                        ),
+                        InlineKeyboardButton(text=constants.MAIN_MENU, callback_data=constants.MENU),
                     ]
                 ],
             ),
         )
 
 
-class StartCallbackQueryEndpoint(CallbackQueryEndpoint):
-    callback_query_name: str = constants.MENU
+class FethiyeCallbackQueryEndpoint(CallbackQueryEndpoint):
+    callback_query_name: str = constants.SOUTH
 
     async def handle(
         self,
@@ -52,18 +53,20 @@ class StartCallbackQueryEndpoint(CallbackQueryEndpoint):
         callback_query: CallbackQuery,
     ) -> None:  # noqa: U100
         await client.answer_callback_query(callback_query.id)
-        template = templates_text.MENU
         await client.send_photo(
             chat_id=callback_query.from_user.id,
-            photo=photo_ids.START,
+            photo=photo_ids.SOUTH,
         )
+        template = templates_text.SOUTH
         await callback_query.message.reply(
             text=template,
             reply_markup=InlineKeyboardMarkup(
                 [
                     [
-                        InlineKeyboardButton(text=constants.GUIDE_SOUTH, callback_data=constants.SOUTH),
-                        InlineKeyboardButton(text=constants.GUIDE_KAS, callback_data=constants.KAS),
+                        InlineKeyboardButton(
+                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.SOUTH}"
+                        ),
+                        InlineKeyboardButton(text=constants.MAIN_MENU, callback_data=constants.MENU),
                     ]
                 ],
             ),
