+++ bot/telegram/commands/start.py
@@ -6,16 +6,17 @@ from pyrogram.types import InlineKeyboardButton
 from pyrogram.types import InlineKeyboardMarkup
 from pyrogram.types import Message
 
-from src.config import constants
-from src.config import photo_ids
-from src.telegram.base import CallbackQueryEndpoint
-from src.telegram.base import PrivateCommandEndpoint
-from src.telegram.templates import text as templates_text
+from bot.config import constants
+from bot.config import photo_ids
+from bot.telegram.base import CallbackQueryEndpoint
+from bot.telegram.base import PrivateCommandEndpoint
+from bot.telegram.templates import text as templates_text
 
 
-class KasEndpoint(PrivateCommandEndpoint):
+class StartEndpoint(PrivateCommandEndpoint):
     commands: List[str] = [
-        constants.KAS,
+        constants.START,
+        constants.MENU,
     ]
 
     async def handle(
@@ -25,27 +26,25 @@ class KasEndpoint(PrivateCommandEndpoint):
     ) -> None:  # noqa: U100
         await client.send_photo(
             chat_id=message.from_user.id,
-            photo=photo_ids.KAS,
+            photo=photo_ids.START,
         )
-        template = templates_text.KAS
+        template = templates_text.MENU
         await client.send_message(
             chat_id=message.from_user.id,
             text=template,
             reply_markup=InlineKeyboardMarkup(
                 [
                     [
-                        InlineKeyboardButton(
-                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.KAS}"
-                        ),
-                        InlineKeyboardButton(text=constants.MAIN_MENU, callback_data=constants.MENU),
+                        InlineKeyboardButton(text=constants.GUIDE_SOUTH, callback_data=constants.SOUTH),
+                        InlineKeyboardButton(text=constants.GUIDE_KAS, callback_data=constants.KAS),
                     ]
                 ],
             ),
         )
 
 
-class KasCallbackQueryEndpoint(CallbackQueryEndpoint):
-    callback_query_name: str = constants.KAS
+class StartCallbackQueryEndpoint(CallbackQueryEndpoint):
+    callback_query_name: str = constants.MENU
 
     async def handle(
         self,
@@ -53,20 +52,18 @@ class KasCallbackQueryEndpoint(CallbackQueryEndpoint):
         callback_query: CallbackQuery,
     ) -> None:  # noqa: U100
         await client.answer_callback_query(callback_query.id)
-        template = templates_text.KAS
+        template = templates_text.MENU
         await client.send_photo(
             chat_id=callback_query.from_user.id,
-            photo=photo_ids.KAS,
+            photo=photo_ids.START,
         )
         await callback_query.message.reply(
             text=template,
             reply_markup=InlineKeyboardMarkup(
                 [
                     [
-                        InlineKeyboardButton(
-                            text=constants.PAYMENT_TEXT, callback_data=f"{constants.PAYMENT}_{constants.KAS}"
-                        ),
-                        InlineKeyboardButton(text=constants.MAIN_MENU, callback_data=constants.MENU),
+                        InlineKeyboardButton(text=constants.GUIDE_SOUTH, callback_data=constants.SOUTH),
+                        InlineKeyboardButton(text=constants.GUIDE_KAS, callback_data=constants.KAS),
                     ]
                 ],
             ),
