+++ bot/telegram/templates/text.py
