+++ compose-dev.yaml
@@ -1,4 +1,5 @@
 version: "3.7"
+
 services:
   bot:
     # build:
@@ -8,8 +9,6 @@ services:
       context: .
       dockerfile: Dockerfile
     restart: unless-stopped
-    secrets:
-      - db-password
     environment:
     - TELEGRAM_ADMIN_ID=${TELEGRAM_ADMIN_ID}
     - TELEGRAM_BOT_NAME=${TELEGRAM_BOT_NAME}
